+++ src/date/util/reset.js
@@ -1,17 +1,14 @@
-import { SPECIFICITY_INDEX, getUnitSpecificity } from './units';
+import { SPECIFICITY_INDEX, getUnitSpecificity, getUnitEdge } from './units';
 import { callDateSet } from './helpers';
 
-export function resetByUnit(date, unit) {
-  return resetBySpecificity(date, getUnitSpecificity(unit));
+export function resetByUnit(date, unit, end) {
+  return resetBySpecificity(date, getUnitSpecificity(unit), end);
 }
 
-export function resetBySpecificity(date, specificity) {
+export function resetBySpecificity(date, specificity, end = false) {
   for (let i = specificity + 1; i < SPECIFICITY_INDEX.length; i++) {
     const unit = SPECIFICITY_INDEX[i];
-    if (unit !== 'week') {
-      // Date reset value is 1, everything else is 0 index.
-      const val = unit === 'date' ? 1 : 0;
-      callDateSet(date, SPECIFICITY_INDEX[i], val);
-    }
+    const val = getUnitEdge(unit, end, date);
+    callDateSet(date, unit, val);
   }
 }